Refactor position of PlayerColorEnum
